+++ NEWS.md
@@ -1,3 +1,7 @@
+# teal.modules.clinical 0.8.4.9000
+
+* issues with no news:
+
 # teal.modules.clinical 0.8.4
 
 * Extend `tm_t_coxreg` to optionally produce univariate Cox regressions.